requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
idna==3.8
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
